refactor(api): move serializer to src
