+++ api/src/shared/infrastructure/serializers/jsonapi/course-serializer.js
