+++ api/tests/shared/unit/infrastructure/serializers/jsonapi/course-serializer_test.js
@@ -1,6 +1,6 @@
-import * as serializer from '../../../../../lib/infrastructure/serializers/jsonapi/course-serializer.js';
-import { Course } from '../../../../../src/shared/domain/models/Course.js';
-import { expect } from '../../../../test-helper.js';
+import { Course } from '../../../../../../src/shared/domain/models/Course.js';
+import * as serializer from '../../../../../../src/shared/infrastructure/serializers/jsonapi/course-serializer.js';
+import { expect } from '../../../../../test-helper.js';
 
 describe('Unit | Serializer | JSONAPI | course-serializer', function () {
   describe('#serialize()', function () {
